requirements.txt
blinker==1.9.0
boto3==1.39.11
botocore==1.39.11
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
google-generativeai>=0.8.5
gspread
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
oauth2client
Pillow>=11.1.0
pillow-heif>=1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.36.0
SQLAlchemy>=1.4.0
psycopg2-binary
urllib3==2.5.0
Werkzeug==3.1.3
